+++ test/extensions/filters/listener/tls_inspector/tls_inspector_test.cc
@@ -104,7 +104,7 @@ TEST_F(TlsInspectorTest, SniRegistered) {
       }));
   EXPECT_CALL(socket_, setRequestedServerName(Eq(servername)));
   EXPECT_CALL(socket_, setRequestedApplicationProtocols(_)).Times(0);
-  EXPECT_CALL(socket_, setDetectedTransportProtocol(absl::string_view("ssl")));
+  EXPECT_CALL(socket_, setDetectedTransportProtocol(absl::string_view("tls")));
   EXPECT_CALL(cb_, continueFilterChain(true));
   file_event_callback_(Event::FileReadyType::Read);
   EXPECT_EQ(1, cfg_->stats().tls_found_.value());
@@ -126,7 +126,7 @@ TEST_F(TlsInspectorTest, AlpnRegistered) {
       }));
   EXPECT_CALL(socket_, setRequestedServerName(_)).Times(0);
   EXPECT_CALL(socket_, setRequestedApplicationProtocols(alpn_protos));
-  EXPECT_CALL(socket_, setDetectedTransportProtocol(absl::string_view("ssl")));
+  EXPECT_CALL(socket_, setDetectedTransportProtocol(absl::string_view("tls")));
   EXPECT_CALL(cb_, continueFilterChain(true));
   file_event_callback_(Event::FileReadyType::Read);
   EXPECT_EQ(1, cfg_->stats().tls_found_.value());
@@ -159,7 +159,7 @@ TEST_F(TlsInspectorTest, MultipleReads) {
   bool got_continue = false;
   EXPECT_CALL(socket_, setRequestedServerName(Eq(servername)));
   EXPECT_CALL(socket_, setRequestedApplicationProtocols(alpn_protos));
-  EXPECT_CALL(socket_, setDetectedTransportProtocol(absl::string_view("ssl")));
+  EXPECT_CALL(socket_, setDetectedTransportProtocol(absl::string_view("tls")));
   EXPECT_CALL(cb_, continueFilterChain(true)).WillOnce(InvokeWithoutArgs([&got_continue]() {
     got_continue = true;
   }));
@@ -183,7 +183,7 @@ TEST_F(TlsInspectorTest, NoExtensions) {
       }));
   EXPECT_CALL(socket_, setRequestedServerName(_)).Times(0);
   EXPECT_CALL(socket_, setRequestedApplicationProtocols(_)).Times(0);
-  EXPECT_CALL(socket_, setDetectedTransportProtocol(absl::string_view("ssl")));
+  EXPECT_CALL(socket_, setDetectedTransportProtocol(absl::string_view("tls")));
   EXPECT_CALL(cb_, continueFilterChain(true));
   file_event_callback_(Event::FileReadyType::Read);
   EXPECT_EQ(1, cfg_->stats().tls_found_.value());