+++ test/mocks/network/mocks.cc
@@ -166,6 +166,12 @@ MockListenerFilterCallbacks::~MockListenerFilterCallbacks() {}
 MockListenerFilterManager::MockListenerFilterManager() {}
 MockListenerFilterManager::~MockListenerFilterManager() {}
 
+MockFilterChain::MockFilterChain() {}
+MockFilterChain::~MockFilterChain() {}
+
+MockFilterChainManager::MockFilterChainManager() {}
+MockFilterChainManager::~MockFilterChainManager() {}
+
 MockFilterChainFactory::MockFilterChainFactory() {
   ON_CALL(*this, createListenerFilterChain(_)).WillByDefault(Return(true));
 }